+++ services/payments/src/simcore_service_payments/rest/routes.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, FastAPI
 
 from .._meta import API_VTAG
-from . import _acks, _auth, _health, _meta
+from . import _acknoledgements, _auth, _health, _meta
 
 
 def setup_rest_api_routes(app: FastAPI):
@@ -10,5 +10,5 @@ def setup_rest_api_routes(app: FastAPI):
     api_router = APIRouter(prefix=f"/{API_VTAG}")
     api_router.include_router(_auth.router, tags=["auth"])
     api_router.include_router(_meta.router, tags=["meta"])
-    api_router.include_router(_acks.router, tags=["acks"])
+    api_router.include_router(_acknoledgements.router, tags=["acks"])
     app.include_router(api_router)
